refactor: split home presentation views
