+++ scripts/season_stage3_ui_unit_check.swift
@@ -28,13 +28,16 @@ let mapView = load("dogArea/Views/MapView/MapView.swift")
 let mapViewModel = load("dogArea/Views/MapView/MapViewModel.swift")
 let mapSettingView = load("dogArea/Views/MapView/MapSubViews/MapSettingView.swift")
 let profileView = load("dogArea/Views/ProfileSettingView/NotificationCenterView.swift")
+let seasonDetailSheetView = load("dogArea/Views/HomeView/HomeSubView/Presentation/HomeSeasonDetailSheetView.swift")
+let seasonResultOverlayView = load("dogArea/Views/HomeView/HomeSubView/Presentation/HomeSeasonResultOverlayView.swift")
 let spec = load("docs/season-stage3-ui-integration-v1.md")
 let report = load("docs/cycle-156-season-stage3-ui-integration-report-2026-03-01.md")
 
-assertTrue(homeView.contains("seasonDetailSheet"), "HomeView should provide season detail sheet")
+assertTrue(homeView.contains("HomeSeasonDetailSheetView"), "HomeView should provide season detail sheet")
 assertTrue(homeView.contains("todayScoreDelta"), "HomeView should show today's season score delta")
-assertTrue(homeView.contains("seasonRewardStatusText"), "HomeView should render reward claim status text")
 assertTrue(homeView.contains("retrySeasonRewardClaim"), "HomeView should expose reward retry action")
+assertTrue(seasonDetailSheetView.contains("seasonDetailLine"), "Season detail sheet view should render season detail rows")
+assertTrue(seasonResultOverlayView.contains("seasonRewardStatusText"), "Season result overlay view should render reward claim status text")
 
 assertTrue(homeViewModel.contains("todayScoreDelta"), "SeasonMotionSummary should include today score delta")
 assertTrue(homeViewModel.contains("dailyScoreLedger"), "SeasonMotionStore should persist daily score ledger")
